pyproject.toml
[tool.poetry]
name = "pdf_portfolio"
version = "0.1.0"
description = ""
authors = ["Artur Kutsevol <arthur.kutsevol@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
django = "^2.2"
django-nested-inline = "^0.3.7"
python-decouple = "^3.1"
django-axes = "^5.0"
invoke = "^1.2"
django-csp = "^3.5"
django-feature-policy = "^2.3"
django-referrer-policy = "^1.0"
whitenoise = "^4.1"
django-split-settings = "^0.3.0"
isort = "^4.3"

[tool.poetry.dev-dependencies]
django-debug-toolbar = "^2.0"
nplusone = "^1.0"
django-querycount = "^0.7.0"
flake8-django = "^0.0.4"
flake8 = "^3.7"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
